fix(image): 이미지 사이즈 소수점 이하 2자리까지 저장

+++ src/main/java/chzzk/grassdiary/domain/image/dto/ImageDTO.java
@@ -8,9 +8,9 @@ public record ImageDTO (
         Long imageId,
         String imageURL,
         String imageName,
-        Long imageSize
+        Double imageSize
 ) {
-    public static ImageDTO from(DiaryToImage diaryToImage, String imageURL, String imageName, Long imageSize) {
+    public static ImageDTO from(DiaryToImage diaryToImage, String imageURL, String imageName, Double imageSize) {
         return new ImageDTO(diaryToImage.getImage().getId(), imageURL, imageName, imageSize);
     }
 }
